requirements.txt
mutagen
pycryptodome

# gui
PyQt6
pyinstaller

# web
streamlit